Fix tests close #450
